requirements.txt
streamlit==1.39.0
python-docx==1.1.2
PyPDF2==3.0.1
sentence-transformers==3.1.1
scikit-learn==1.5.1
numpy==1.26.4
razorpay==1.4.1
python-dotenv==1.0.1
pandas==2.2.2
fpdf2==2.7.9  